requirements.txt
python-telegram-bot==13.15
tweepy
requests
python-dotenv
yt-dlp